requirements.txt
google-generativeai>=0.3.0
google-genai
python-dotenv>=0.21.0
rich>=13.4.2
prompt-toolkit>=3.0.0
numpy>=1.22.0
